Reduce build-deps for "make dist-xz" (getgit:claws)
Set build options, so "make dist-xz" doesn't depend on gtk (and other
libraries).

Creating a tar package from a git checkout of claws-mail is currently
required, as the plugins and docs expect the generated sources in source
dir (which are generated in build dir, thus not found). It also
generates the version numbers as required (affects file "version" and
directory- and archive-name).